♻️ Split stats
