+++ src/HonzaBotner.Discord.Services/Jobs/TriggerRemindersJobProvider.cs
@@ -44,7 +44,7 @@ public class TriggerRemindersJobProvider : IJob
         _reminderManager = reminderManager;
     }
 
-    public string Name { get; } = "reminders-trigger";
+    public string Name => "reminders-trigger";
 
     public async Task ExecuteAsync(CancellationToken cancellationToken)
     {
